# Conflicts:
#	src/main/java/ac/dankook/dht2/controller/AverageDataRestController.java
#	src/main/java/ac/dankook/dht2/data/AverageData.java
#	src/main/java/ac/dankook/dht2/repository/AverageRepository.java
#	src/main/java/ac/dankook/dht2/service/AverageService.java
#	src/main/java/ac/dankook/dht2/service/AverageServiceImpl.java
#	src/main/resources/mybatis/data-mapper.xml
#	src/main/webapp/resources/js/average.js

+++ src/main/java/ac/dankook/dht2/data/Data.java
@@ -1,19 +1,19 @@
 package ac.dankook.dht2.data;
 
 public class Data {
-    private short temperature;
-    private short humidity;
+    private int temperature;
+    private int humidity;
     private String user_id;
 
     public String getUser_id() {return user_id;}
 
     public void setUser_id(String user_id) {this.user_id = user_id;}
 
-    public short getTemperature() {return temperature;}
+    public int getTemperature() {return temperature;}
 
-    public void setTemperature(short temperature) {this.temperature = temperature;}
+    public void setTemperature(int temperature) {this.temperature = temperature;}
 
-    public short getHumidity() {return humidity;}
+    public int getHumidity() {return humidity;}
 
-    public void setHumidity(short humidity) {this.humidity = humidity;}
+    public void setHumidity(int humidity) {this.humidity = humidity;}
 }
